Modified type in temp_kf because of not vaild type

+++ app/src/main/java/ca/nismit/util/weather/pojoForecast/Main.java
@@ -29,7 +29,7 @@ public class Main {
     private Integer humidity;
     @SerializedName("temp_kf")
     @Expose
-    private Integer tempKf;
+    private Float tempKf;
 
     public Double getTemp() {
         return temp;
@@ -87,11 +87,11 @@ public class Main {
         this.humidity = humidity;
     }
 
-    public Integer getTempKf() {
+    public Float getTempKf() {
         return tempKf;
     }
 
-    public void setTempKf(Integer tempKf) {
+    public void setTempKf(Float tempKf) {
         this.tempKf = tempKf;
     }
